+++ src/proto/media/file/info.rs
@@ -41,7 +41,7 @@ trait InfoInternal: Align + ptr::Pointee<Metadata = usize> {
     /// struct.
     unsafe fn name_ptr(ptr: *mut u8) -> *mut Char16 {
         let offset_of_str = Self::name_offset();
-        ptr.add(offset_of_str) as *mut Char16
+        ptr.add(offset_of_str).cast::<Char16>()
     }
 
     /// Create a new info type in user-provided storage.
@@ -83,17 +83,18 @@ trait InfoInternal: Align + ptr::Pointee<Metadata = usize> {
 
         // Create a raw fat pointer using the `storage` as a base.
         let info_ptr: *mut Self =
-            ptr::from_raw_parts_mut(storage.as_mut_ptr() as *mut (), name_length_ucs2);
+            ptr::from_raw_parts_mut(storage.as_mut_ptr().cast::<()>(), name_length_ucs2);
 
         // Initialize the struct header.
         init(info_ptr, info_size as u64);
 
+        // Create a pointer to the part of info where the name is
+        // stored. Note that `info_ptr` is used rather than `storage` to
+        // comply with Stacked Borrows.
+        let info_name_ptr = Self::name_ptr(info_ptr.cast::<u8>());
+
         // Initialize the name slice.
-        ptr::copy(
-            name.as_ptr(),
-            Self::name_ptr(storage.as_mut_ptr()),
-            name_length_ucs2,
-        );
+        ptr::copy(name.as_ptr(), info_name_ptr, name_length_ucs2);
 
         // The struct is now valid and safe to dereference.
         let info = &mut *info_ptr;
@@ -106,10 +107,10 @@ where
     T: InfoInternal + ?Sized,
 {
     unsafe fn from_uefi<'ptr>(ptr: *mut c_void) -> &'ptr mut Self {
-        let name_ptr = Self::name_ptr(ptr as *mut u8);
+        let name_ptr = Self::name_ptr(ptr.cast::<u8>());
         let name = CStr16::from_ptr(name_ptr);
         let name_len = name.as_slice_with_nul().len();
-        &mut *ptr::from_raw_parts_mut(ptr as *mut (), name_len)
+        &mut *ptr::from_raw_parts_mut(ptr.cast::<()>(), name_len)
     }
 }